pyproject.toml
[tool.poetry]
name = "campaign-pipeline"
version = "0.1.0"
description = "Pipeline for FEC and vote alignment"
authors = ["AutoDev <auto@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
sqlmodel = "^0.0.14"
pandas = "^2.2.1"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^1.10.9"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
duckdb = "^0.9.2"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
campaign-cli = "src.cli:app"
